feat: restructure episode endpoints into v1 router
